pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-retrieval-mcp"
version = "0.1.0"
description = "Lightweight MCP Server for RAG-based skill retrieval"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    "click>=8.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
openai = ["openai>=1.50", "tiktoken>=0.7"]
local = ["sentence-transformers>=3.0"]
ollama = ["httpx>=0.27"]
all = ["openai>=1.50", "tiktoken>=0.7", "sentence-transformers>=3.0", "httpx>=0.27"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5"]

[project.scripts]
skill-mcp = "skill_mcp.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"
